refactored to make example easier to build and understand
