+++ examples/slack-echo-flow/.dockerignore
@@ -1,1 +1,2 @@
+deployments
 slack-echo-flow
