+++ examples/slack-echo-flow/Dockerfile
@@ -1,12 +1,12 @@
 #Img build
 FROM golang:1.11 AS builder
-WORKDIR /echo-nats
+WORKDIR /slack-echo-flow
 COPY . ./
 RUN go test -v -mod=readonly ./...
 RUN CGO_ENABLED=0 go build -mod=readonly
 
 # Compile statically and use scratch (blank) image to reduce img size
 FROM alpine:latest
-WORKDIR /echo-nats
-COPY --from=builder /echo-nats/echo-nats .
-CMD ["./echo-nats"]
+WORKDIR /slack-echo-flow
+COPY --from=builder /slack-echo-flow/slack-echo-flow .
+CMD ["./slack-echo-flow"]
