+++ examples/slack-echo-flow/go.mod
@@ -1,4 +1,4 @@
-module echo-nats
+module slack-echo-flow
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
